test_requirements.txt
# NB - this is the testing requirements - it omits some packages which are not covered by the tests
# Some packages are pinned to old versions to be sure that the code runs against the old package versions (assuming forwards compatibility
# will also let us use more recent packages). One of the main drivers for testing so far back is to be able to roll bugfixes
# out rapidly to users where completely installing from scratch is not an option.
# TODO - also test against bleeding edge (but only after we have current tests working reliably)

numpy==1.14.*
numexpr<=2.7.1 #2.7.2 conflicts with numpy
scipy
matplotlib<=3.2.* # matplotlib >=3.3.0 requires numpy >=1.15 (and pip isn't sensible enough to recognise this)
#wxpython
tables #<=3.4.2
#pyopengl
traits
traitsui
pillow
cython

toposort
#networkx

#pyro

#pyfftw3; python_version < '3.0'
pyfftw #; python_version > '3.0'

#mpld3
cherrypy
scikit-image
scikit-learn

#FIXME for py3k
#shapely [osx and py2k]

zeroconf==0.17.7
#zeroconf
requests
#pandas
pyyaml
psutil

#docutils
#sphinx
ujson
jinja2
#dispatch
#pycairo
#pymecompress >=0.2
six
future
#pywin32 [win]
cython
